requirements.txt
asgiref==3.2.3
Django==3.0.1
django-mdeditor==0.1.17
Markdown==3.1.1
mistletoe==0.7.2
mysqlclient==1.4.6
Pillow==6.2.1
Pygments==2.5.2
PyMySQL==0.9.3
pytz==2019.3
sqlparse==0.3.0
